backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.31.0
python-multipart>=0.0.12
sqlmodel>=0.0.21
google-generativeai>=0.8.0
pdfplumber>=0.11.0
python-docx>=1.1.0
PyMuPDF>=1.24.0
numpy>=1.26.0
python-dotenv>=1.0.0
pillow>=10.0.0
